Update readme to include both node.js and HTML/JS
This can be used both on the server side (Node.js) and the client side (HTML/JS). This is a requirement because the search engine can also even be run on local disk using URI or as a global server. Keeping in with the flexible usability of the overall smart contract search engine product.